+++ lib/screens/dashboard/wallets/wallet_item.dart
@@ -8,7 +8,7 @@ import 'package:provenance_wallet/extension/coin_extension.dart';
 import 'package:provenance_wallet/screens/dashboard/dashboard_bloc.dart';
 import 'package:provenance_wallet/screens/dashboard/wallets/rename_wallet_dialog.dart';
 import 'package:provenance_wallet/screens/dashboard/wallets/wallets_bloc.dart';
-import 'package:provenance_wallet/services/key_value_service.dart';
+import 'package:provenance_wallet/services/key_value_service/key_value_service.dart';
 import 'package:provenance_wallet/services/models/wallet_details.dart';
 import 'package:provenance_wallet/services/wallet_service/wallet_service.dart';
 import 'package:provenance_wallet/util/get.dart';
