+++ lib/services/key_value_service/shared_preferences_key_value_store.dart
@@ -1,6 +1,6 @@
 import 'dart:async';
 
-import 'package:provenance_wallet/services/key_value_service.dart';
+import 'package:provenance_wallet/services/key_value_service/key_value_service.dart';
 import 'package:provenance_wallet/services/key_value_service/key_value_store.dart';
 import 'package:shared_preferences/shared_preferences.dart';
